+++ example/shared/src/components/Header.tsx
@@ -16,8 +16,8 @@ export const Header = ({ children, style, ...rest }: HeaderProps) => {
 
 const styles = StyleSheet.create({
   container: {
-    height: Platform.select({ android: HEADER_HEIGHT + SPACING, default: HEADER_HEIGHT }),
-    paddingTop: Platform.select({ android: SPACING * 2 }),
+    height: Platform.select({ ios: HEADER_HEIGHT, default: HEADER_HEIGHT + SPACING }),
+    paddingTop: Platform.select({ ios: 0, default: SPACING * 2 }),
     justifyContent: 'center',
     padding: SPACING,
   },